% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faInit.R
\name{faInit}
\alias{faInit}
\title{Generate initial parameters for EM/FA}
\usage{
faInit(X_t_unsupervised, nDimZ, type = "fa", checkArgs = TRUE)
}
\arguments{
\item{X_t_unsupervised}{train input data as design matrix (must be matrix, not data.frame), i.e. row = instance, column = feature/attribute}

\item{nDimZ}{how many latent factors to consider}

\item{type}{"unconstrained", "fa", or "ppca"; refers to psi}

\item{checkArgs}{whether to check the arguments are valid; it takes more time to execute}
}
\value{
initial parameters for EM/FA, i.e. a list containing nDimX, nDimZ, type, mu_t, lambda_t, psi_t. "_t" comes from "transpose"
}
\description{
Generate initial parameters for EM/FA
}
\examples{
params <- faInit(X_t_unsupervised = house[,-1,drop=FALSE],
                 nDimZ=1,
                 type = "fa",checkArgs = TRUE)
params
}
