% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faPlot.R
\name{faPlot}
\alias{faPlot}
\title{Plot the resulted FA hyperplane}
\usage{
faPlot(X_t, Z_t, p, color = "red", add = FALSE, addPoints = FALSE,
  pointsColor = "black", addLr = FALSE, lrColor = "blue",
  checkArgs = TRUE, checkPositiveDefinite = FALSE)
}
\arguments{
\item{X_t}{a matrix representing the input points to be plotted (if addPoints is TRUE); usually, the training input}

\item{Z_t}{a matrix representing the output = values of function in input points; usually, the training output}

\item{p}{parameters of a fitted FA}

\item{color}{color of the hyperplane given by FA}

\item{add}{add this plot to an already existing plot?}

\item{addPoints}{whether to plot the points X_t, Z_t}

\item{pointsColor}{color of points}

\item{addLr}{fit and add a linear regression model on plot?}

\item{lrColor}{color of linear regression hyperplane}

\item{checkArgs}{whether to check the arguments are valid; it takes more time to execute}

\item{checkPositiveDefinite}{whether to check the covariance matrices are valid; it takes more time to execute}
}
\description{
Works only if input is 1- or 2-dimensional and output is 1-dimensional.
}
\examples{
params <- faFit(X_t_unsupervised = house[,2:3,drop=FALSE],
                paramsNULL_nDimZ = 1,
                params=NULL,
                type="fa",
                lambdaRidge=0,
                checkArgs=FALSE,
                checkPositiveDefinite=FALSE,
                epsilon=1e-10,
                maxIterations=100,
                stopType="parameter",
                turboEmMethods=c("em","pem"))
faPlot(X_t=house[,2:3,drop=FALSE],
       Z_t=house[,1,drop=FALSE],
       p=params,
       color="red",
       add=FALSE,
       addPoints=TRUE,
       pointsColor="black",
       addLr=TRUE,
       lrColor="blue",
       checkArgs=TRUE,
       checkPositiveDefinite=FALSE)

params <- faFit(X_t_unsupervised = house[,2,drop=FALSE],
                paramsNULL_nDimZ = 1,
                params=NULL,
                type="fa",
                lambdaRidge=0,
                checkArgs=FALSE,
                checkPositiveDefinite=FALSE,
                epsilon=1e-10,
                maxIterations=100,
                stopType="parameter",
                turboEmMethods=c("em","pem"))
faPlot(X_t=house[,2,drop=FALSE],
       Z_t=house[,1,drop=FALSE],
       p=params,
       color="red",
       add=FALSE,
       addPoints=TRUE,
       pointsColor="black",
       addLr=TRUE,
       lrColor="blue",
       checkArgs=TRUE,
       checkPositiveDefinite=FALSE)
}
