% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faPredict.R
\name{faPredict}
\alias{faPredict}
\title{Predict new values, given the parameters of a trained FA model}
\usage{
faPredict(p, X_t_test, checkArgs = TRUE, checkPositiveDefinite = FALSE)
}
\arguments{
\item{p}{parameters of a fitted FA}

\item{X_t_test}{a matrix containing the input data; their output values would be predicted}

\item{checkArgs}{whether to check the arguments are valid; it takes more time to execute}

\item{checkPositiveDefinite}{whether to check the covariance matrices are valid; it takes more time to execute}
}
\value{
a list containing the transposed predicted values ($values_t) and the covariance matrix associated with them ($Sigma_t)
}
\description{
Predict new values, given the parameters of a trained FA model
}
\examples{
params <- faFit(X_t_unsupervised = house[1:20,2:3,drop=FALSE],
                paramsNULL_nDimZ = 2,
                params=NULL,
                type="fa",
                lambdaRidge=0,
                checkArgs=FALSE,
                checkPositiveDefinite=FALSE,
                epsilon=1e-10,
                maxIterations=100,
                stopType="parameter",
                turboEmMethods=NULL)
result <- faPredict(p = params,
                    X_t_test = house[21:26,2:3,drop=FALSE],
                    checkArgs=TRUE,
                    checkPositiveDefinite=FALSE)
result

}
