% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mS3faImpute.R
\name{mS3faImpute}
\alias{mS3faImpute}
\title{Impute missing data via MS3FA}
\usage{
mS3faImpute(params, testInput, testOutput)
}
\arguments{
\item{params}{parameters of a fitted MS3FA}

\item{testInput}{test input data; can contain NAs; error if a full row (input+output) is NA}

\item{testOutput}{train output data; can contain NAs; error if a full row (input+output) is NA}
}
\value{
a list with two elements: testInput with NAs imputed, testOutput with NAs imputed
}
\description{
Impute missing data via MS3FA
}
\examples{
houseCopy <- house
houseCopy[1:10,2:3] <- NA
houseCopy[11:20,1] <- NA
params <- mS3faFit(trainInput=houseCopy[,2:3,drop=FALSE],
                   trainOutput=houseCopy[,1,drop=FALSE],
                   params=NULL,
                   type="ppca",
                   lambdaRidge=0,
                   epsilon=1e-10,
                   maxIterations=100,
                   stopType="objfn",
                   turboEmMethods=NULL)
result <- mS3faImpute(params=params,
                      testInput=houseCopy[,2:3,drop=FALSE],
                      testOutput=houseCopy[,1,drop=FALSE])
result
}
