% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mS3faInit.R
\name{mS3faInit}
\alias{mS3faInit}
\title{Generate initial parameters for EM/MS3FA}
\usage{
mS3faInit(trainInput, trainOutput, type = "fa")
}
\arguments{
\item{trainInput}{train input data; can contain NAs; error if a full row (input+output) is NA}

\item{trainOutput}{train output data; can contain NAs; error if a full row (input+output) is NA}

\item{type}{"unconstrained", "fa", or "ppca"; refers to psi}
}
\value{
initial parameters for EM/MS3FA, i.e. a list containing nDimX, nDimZ, type, mu_z_t, Sigma_z_t, mu_t, lambda_t, psi_t. "_t" comes from "transpose"
}
\description{
Generate initial parameters for EM/MS3FA
}
\examples{
houseCopy <- house
house[1:10,2:3] <- NA
house[11:20,1] <- NA
params0 <- mS3faInit(trainInput=houseCopy[,2:3,drop=FALSE],
                     trainOutput=houseCopy[,1,drop=FALSE],
                     type="fa")
}
