% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mS3faPredict.R
\name{mS3faPredict}
\alias{mS3faPredict}
\title{Predict new values, given the parameters of a trained MS3FA model}
\usage{
mS3faPredict(params, testInput, testOutput)
}
\arguments{
\item{params}{parameters of a fitted MS3FA}

\item{testInput}{test input data; can contain NAs; error if a full row (input+output) is NA}

\item{testOutput}{train output data; can contain NAs; error if a full row (input+output) is NA}
}
\value{
a list with two elements: '$values' only with imputed NAs, '$Sigmas' with the covariance matrices associated with them (one for each)
}
\description{
Predict new values, given the parameters of a trained MS3FA model
}
\examples{
houseCopy <- house
houseCopy[1:10,2:3] <- NA
houseCopy[11:20,1] <- NA
params <- mS3faFit(trainInput=houseCopy[,2:3,drop=FALSE],
                   trainOutput=houseCopy[,1,drop=FALSE],
                   params=NULL,
                   type="ppca",
                   lambdaRidge=0,
                   epsilon=1e-10,
                   maxIterations=100,
                   stopType="objfn",
                   turboEmMethods=NULL)
result <- mS3faPredict(params=params,
                       testInput=houseCopy[,2:3,drop=FALSE],
                       testOutput=houseCopy[,1,drop=FALSE])
result
}
