% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2faFit.R
\name{s2faFit}
\alias{s2faFit}
\title{Fit S2FA to data via analytic solution}
\usage{
s2faFit(X_t, Z_t, type = "fa", lambdaRidge = 0, checkArgs = TRUE)
}
\arguments{
\item{X_t}{train input data as design matrix (must be matrix, not data.frame), i.e. row = instance, column = feature/attribute}

\item{Z_t}{train output data as design matrix (must be matrix, not data.frame), i.e. row = instance, column = feature/attribute}

\item{type}{"unconstrained", "fa", or "ppca"; refers to psi}

\item{lambdaRidge}{L2 regularization term; must be a number}

\item{checkArgs}{whether to check the arguments are valid; it takes more time to execute}
}
\value{
final parameters learnt, i.e. a list containing nDimX, nDimZ, type, mu_z_,t Sigma_z_t, mu_t, lambda_t, psi_t. "_t" comes from "transpose"
}
\description{
Fit S2FA to data via analytic solution
}
\examples{
params <- s2faFit(X_t=house[,2:3,drop=FALSE],
                  Z_t=house[,1,drop=FALSE],
                  type = "fa",
                  lambdaRidge=0,
                  checkArgs=FALSE)
params
}
