% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2faPredict.R
\name{s2faPredict}
\alias{s2faPredict}
\title{Predict new values, given the parameters of a trained S2FA model}
\usage{
s2faPredict(p, X_t_test, checkArgs = TRUE,
  checkPositiveDefinite = FALSE)
}
\arguments{
\item{p}{parameters of a fitted S2FA}

\item{X_t_test}{a matrix containing the input data; their output values would be predicted}

\item{checkArgs}{whether to check the arguments are valid; it takes more time to execute}

\item{checkPositiveDefinite}{whether to check the covariance matrices are valid; it takes more time to execute}
}
\value{
a list containing the transposed predicted values ($values_t) and the covariance matrix associated with them ($Sigma_t)
}
\description{
Predict new values, given the parameters of a trained S2FA model
}
\examples{
params <- s2faFit(X_t=house[,2:3,drop=FALSE],
                  Z_t=house[,1,drop=FALSE],
                  type = "fa",
                  lambdaRidge=0,
                  checkArgs=FALSE)
result <- s2faPredict(p = params,
                    X_t_test = house[21:26,2:3,drop=FALSE],
                    checkArgs=TRUE,
                    checkPositiveDefinite=FALSE)
result
}
