% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3faInit.R
\name{s3faInit}
\alias{s3faInit}
\title{Generate initial parameters for EM/S3FA}
\usage{
s3faInit(X_t_supervised, Z_t_supervised, X_t_unsupervised, type = "fa",
  checkArgs = TRUE)
}
\arguments{
\item{X_t_supervised}{train input data (which has output) as design matrix (must be matrix, not data.frame), i.e. row = instance, column = feature/attribute}

\item{Z_t_supervised}{train output data (for X_t_supervised) as design matrix (must be matrix, not data.frame), i.e. row = instance, column = feature/attribute}

\item{X_t_unsupervised}{train input data (which has no output) as design matrix (must be matrix, not data.frame), i.e. row = instance, column = feature/attribute}

\item{type}{"unconstrained", "fa", or "ppca"; refers to psi}

\item{checkArgs}{whether to check the arguments are valid; it takes more time to execute}
}
\value{
initial parameters for EM/S3FA, i.e. a list containing nDimX, nDimZ, type, mu_z_t, Sigma_z_t, mu_t, lambda_t, psi_t. "_t" comes from "transpose"
}
\description{
Generate initial parameters for EM/S3FA
}
\examples{
params <- s3faInit(X_t_supervised = house[1:10,2:3,drop=FALSE],
                   Z_t_supervised = house[1:10,1,drop=FALSE],
                   X_t_unsupervised = house[11:20,2:3,drop=FALSE],
                   type = "fa",
                   checkArgs = TRUE)
params
}
