% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3faPredict.R
\name{s3faPredict}
\alias{s3faPredict}
\title{Predict new values, given the parameters of a trained S3FA model}
\usage{
s3faPredict(p, X_t_test, checkArgs = TRUE,
  checkPositiveDefinite = FALSE)
}
\arguments{
\item{p}{parameters of a fitted S3FA}

\item{X_t_test}{a matrix containing the input data; their output values would be predicted}

\item{checkArgs}{whether to check the arguments are valid; it takes more time to execute}

\item{checkPositiveDefinite}{whether to check the covariance matrices are valid; it takes more time to execute}
}
\value{
a list containing the transposed predicted values ($values_t) and the covariance matrix associated with them ($Sigma_t)
}
\description{
Predict new values, given the parameters of a trained S3FA model
}
\examples{
params <- s3faFit(X_t_supervised = house[1:10,2:3,drop=FALSE],
                  Z_t_supervised = house[1:10,1,drop=FALSE],
                  X_t_unsupervised = house[11:20,2:3,drop=FALSE],
                  params=NULL,
                  type="fa",
                  lambdaRidge=0,
                  checkArgs=FALSE,
                  checkPositiveDefinite=FALSE,
                  epsilon=1e-10,
                  maxIterations=100,
                  stopType="objfn",
                  turboEmMethods=NULL)
result <- s3faPredict(p = params,
                      X_t_test = house[11:20,2:3,drop=FALSE],
                      checkArgs=TRUE,
                      checkPositiveDefinite=FALSE)
result
}
